Added game and battle managers in separate scene with basic (but incomplete) functionality
Both managers are in a scene labeled "Managers".  The game manager invokes the Don'tDestroyOnLoad method which allows it to persist between scenes.  As of right now, both managers have events that track when a state is changed.  In the battle manager, the timer, start and end actions are not yet implemented.